+++ test/index.js
@@ -10,6 +10,10 @@ var postcss = PostCSS()
           throw new Error("error detected: " + decl.value)
         }
 
+        if (decl.value.indexOf("multierr(") > -1) {
+          throw new Error("error detected: \n" + decl.value)
+        }
+
         return "world"
       }, decl.source)
     })
@@ -22,6 +26,8 @@ test("postcss try helper", function(t) {
   }
   catch (e) {
     var stack = e.stack.split("\n")
+    t.equal(e.message, process.cwd() + "/file.css:2:6: error detected: error(alert!)", "try() should throw an exception that contains a enhanced message")
+    t.equal(e.originalMessage, "error detected: error(alert!)", "try() should throw an exception that contains the original message as a property")
     t.equal(stack[0], "Error: " + process.cwd() + "/file.css:2:6: error detected: error(alert!)", "try() should have an explicit message (sourcefile:lineno:column: message)")
     t.equal(stack[1], "    at " + process.cwd() + "/file.css:2:6", "try() should have a new item in the stack trace (sourcefile:lineno:column)")
   }
@@ -34,6 +40,16 @@ test("postcss try helper", function(t) {
     var stack = e.stack.split("\n")
     t.equal(stack[0], "Error: <css input>:2:6: error detected: error(alert!)", "try() should have an explicit message (<css input>:lineno:column: message)")
     t.equal(stack[1], "    at <css input>:2:6", "try() should have a new item in the stack trace (<css input>:lineno:column)")
+  }
+
+  try {
+    postcss.process("\nthis{throws:multierr(alert!)}")
+    t.fail("should throw an error with adjusted stack")
+  }
+  catch (e) {
+    var stack = e.stack.split("\n")
+    t.equal(stack[0], "Error: <css input>:2:6: error detected: ", "try() should have an explicit message (<css input>:lineno:column: message)")
+    t.equal(stack[2], "    at <css input>:2:6", "try() should have a new item in the stack trace (<css input>:lineno:column)")
   }
 
   t.equal(postcss.process("say{hello:people}").css, "say{hello:world}", "try() should return the value of the callback if nothing happen")