+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.1 - 2014-11-24
+
+- Fix issue with multilines error message in stack trace
+- Add `originalMessage` property in the exception
+
 # 1.1.0 - 2014-11-24
 
 - `try` now returns the result of the callback